pyproject.toml
[project]
name = "hiro"
version = "0.1.0"
description = "hiro — voice-locked agentic assistant (Claude + Resemblyzer speaker ID + Groq Whisper)"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # LLM
    "anthropic>=0.40.0",
    # Audio capture + VAD
    "pyaudio>=0.2.14",
    "webrtcvad>=2.0.10",
    # Speaker verification (SpeechBrain ECAPA-TDNN)
    "speechbrain>=1.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    # STT (Groq Whisper)
    "groq>=1.0.0",
    # TTS (offline fallback)
    "pyttsx3>=2.90",
    # Core
    "numpy>=1.26.0,<2",
    # Skills
    "requests>=2.31.0",
    "finnhub-python>=2.4.0",
    "spotipy>=2.23.0",
    "paho-mqtt>=2.0.0",
    # Config
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "speedtest-cli>=2.1.3",
]

[project.scripts]
hiro = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.0.0"]

[tool.uv]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }

[tool.hatch.build.targets.wheel]
include = ["*.py", "skills/**"]
